+++ src/int_utils.h
@@ -24,8 +24,8 @@
 // fallback because c++11 functions are not available
 namespace fallback {
 namespace std {
-long long int stoll(const ::std::string &, const int base = 10);
-unsigned long long int stoull(const ::std::string &, const int base = 10);
+int64_t stoll(const ::std::string &, const int base = 10);
+uint64_t stoull(const ::std::string &, const int base = 10);
 }
 }
 
@@ -47,14 +47,14 @@ bool within_bounds(const T &val) noexcept {
 /** convert to signed types */
 template <typename T,
           typename std::enable_if<std::is_signed<T>::value>::type * = nullptr>
-long long int str_to_integral_helper(const std::string &string) {
+int64_t str_to_integral_helper(const std::string &string) {
   return fallback::std::stoll(string);
 }
 
 /** convert to unsigned types */
 template <typename T,
           typename std::enable_if<std::is_unsigned<T>::value>::type * = nullptr>
-unsigned long long int str_to_integral_helper(const std::string &string) {
+uint64_t str_to_integral_helper(const std::string &string) {
   return fallback::std::stoull(string);
 }
