+++ tests/int_utils_test.cpp
@@ -45,42 +45,40 @@ public:
   void outofposbounds() { str_to_integral<uint8_t>("256"); }
 
   void test_stoll() {
-    CPPUNIT_ASSERT_EQUAL(static_cast<long long int>(123),
+    CPPUNIT_ASSERT_EQUAL(static_cast<int64_t>(123),
                          fallback::std::stoll("123"));
-    CPPUNIT_ASSERT_EQUAL(static_cast<long long int>(-123),
+    CPPUNIT_ASSERT_EQUAL(static_cast<int64_t>(-123),
                          fallback::std::stoll("-123"));
-    CPPUNIT_ASSERT_EQUAL(static_cast<long long int>(0),
-                         fallback::std::stoll("0"));
+    CPPUNIT_ASSERT_EQUAL(static_cast<int64_t>(0), fallback::std::stoll("0"));
     CPPUNIT_ASSERT_THROW(fallback::std::stoll("1234567890123456789123456789"),
                          std::out_of_range);
     CPPUNIT_ASSERT_THROW(fallback::std::stoll("-1234567890123456789123456789"),
                          std::out_of_range);
     CPPUNIT_ASSERT_THROW(fallback::std::stoll("fdasfd"), std::invalid_argument);
 
-    CPPUNIT_ASSERT_EQUAL(static_cast<long long int>(0),
+    CPPUNIT_ASSERT_EQUAL(static_cast<int64_t>(0),
                          fallback::std::stoll("0", 16));
-    CPPUNIT_ASSERT_EQUAL(static_cast<long long int>(10),
+    CPPUNIT_ASSERT_EQUAL(static_cast<int64_t>(10),
                          fallback::std::stoll("a", 16));
-    CPPUNIT_ASSERT_EQUAL(static_cast<long long int>(15),
+    CPPUNIT_ASSERT_EQUAL(static_cast<int64_t>(15),
                          fallback::std::stoll("F", 16));
   }
 
   void test_stoull() {
-    CPPUNIT_ASSERT_EQUAL(static_cast<unsigned long long int>(123),
+    CPPUNIT_ASSERT_EQUAL(static_cast<uint64_t>(123),
                          fallback::std::stoull("123"));
-    CPPUNIT_ASSERT_EQUAL(static_cast<unsigned long long int>(0),
-                         fallback::std::stoull("0"));
+    CPPUNIT_ASSERT_EQUAL(static_cast<uint64_t>(0), fallback::std::stoull("0"));
     CPPUNIT_ASSERT_THROW(fallback::std::stoull("-123"), std::out_of_range);
     CPPUNIT_ASSERT_THROW(fallback::std::stoull("12345678901234567890123456789"),
                          std::out_of_range);
     CPPUNIT_ASSERT_THROW(fallback::std::stoull("fdasfd"),
                          std::invalid_argument);
 
-    CPPUNIT_ASSERT_EQUAL(static_cast<unsigned long long int>(0),
+    CPPUNIT_ASSERT_EQUAL(static_cast<uint64_t>(0),
                          fallback::std::stoull("0", 16));
-    CPPUNIT_ASSERT_EQUAL(static_cast<unsigned long long int>(10),
+    CPPUNIT_ASSERT_EQUAL(static_cast<uint64_t>(10),
                          fallback::std::stoull("a", 16));
-    CPPUNIT_ASSERT_EQUAL(static_cast<unsigned long long int>(15),
+    CPPUNIT_ASSERT_EQUAL(static_cast<uint64_t>(15),
                          fallback::std::stoull("F", 16));
   }
